spaces/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pymongo>=4.8.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-mongodb>=0.1.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-classic>=1.0.0
langsmith>=0.1.0
langchain-huggingface>=0.1.0
rank-bm25>=0.2.2
pypdf>=5.0.0
langchain-text-splitters>=0.3.0
nest-asyncio>=1.6.0
llama-parse>=0.5.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
